refreactoring removing the label
