+++ .vscode/launch.json
@@ -31,6 +31,22 @@
             ]
         },
         {
+            "name": "Python: Debug Wgsa add",
+            "type": "python",
+            "request": "launch",
+            "module": "wgsa_add.add_annotations",
+            "args": [
+                "-f",  "./resources/test_wgsa_add/input/chr21.vcf",
+                "-p",  "./resources/panther",
+                "-e",  "./../annoq-data/enhancer/enhancer_map"
+            ],
+            "console": "integratedTerminal",
+            "cwd": "${workspaceFolder}",
+            "env": {
+                "PYTHONPATH": "${workspaceFolder}"
+            }
+        },
+        {
             "name": "Python: Depickle",
             "type": "python",
             "request": "launch",
