+++ wgsa_add/add_annotations.py
@@ -1,12 +1,12 @@
 import argparse
-from collections import defaultdict
-from utils import *
-from base import load_json, load_pickle
 from os import path as ospath
-from add_panther_anno import add_annotation_header as add_panther_annotation_header
-from add_panther_anno import add_annotation_row as add_panther_annotation_row
-from add_enhancer_anno import add_annotation_header as add_enhancer_annotation_header
-from add_enhancer_anno import add_annotation_row as add_enhancer_annotation_row
+from collections import defaultdict
+from wgsa_add.utils import *
+from wgsa_add.base import load_json, load_pickle
+from wgsa_add.add_panther_anno import add_annotation_header as add_panther_annotation_header
+from wgsa_add.add_panther_anno import add_annotation_row as add_panther_annotation_row
+from wgsa_add.add_enhancer_anno import add_annotation_header as add_enhancer_annotation_header
+from wgsa_add.add_enhancer_anno import add_annotation_row as add_enhancer_annotation_row
 
 
 def nested_dict(): return defaultdict(nested_dict)
@@ -28,8 +28,7 @@ def main():
     gene_coords = {i[2]: (i[1][1], i[1][2]) for i in coord_data}
     # [ENSG_id, (contig, start, end), HGNC_id]
 
-    add_annotations(vcf_path, annotation, annoq_tree,
-                    panther_data, gene_coords, lambda x: print(x.rstrip()))
+    add_annotations(vcf_path, annotation, annoq_tree, panther_data, gene_coords)
 
 
 def parse_arguments():
@@ -55,7 +54,7 @@ def add_annotations(filepath, annotations, annoq_tree, panther_data, gene_coords
                                              tool_idxs=tool_idxs)
         cols += add_enhancer_annotation_header()
 
-        print(add_record(row, cols))
+        print_line(add_record(row, cols))
         # add info
         while row:
             row = fp.readline().rstrip()
@@ -64,8 +63,11 @@ def add_annotations(filepath, annotations, annoq_tree, panther_data, gene_coords
                                                   tool_idxs=tool_idxs)
                 cols += add_enhancer_annotation_row( row, annotations)
 
-                print(add_record(row, cols))
+                print_line(add_record(row, cols))
 
 
 if __name__ == "__main__":
     main()
+
+
+#python3 -m wgsa_add.add_annotations -f ./resources/test_wgsa_add/input/chr2.vcf -p resources/panther -e ./../annoq-data/enhancer/enhancer_map/ > resources/test_wgsa_add/output/chr2.vcf
