+++ wgsa_add/hrc_batch.template
@@ -10,15 +10,15 @@
 
 in_file='$IN_FILE'
 out_file='$OUT_FILE'
-panther_dir='./../../annoq-data/wgsa_add'
-enhancer_dir='./../../annoq-data/enhancer/enhancer_map'
+panther_dir='./../annoq-data/wgsa_add'
+enhancer_dir='./../annoq-data/enhancer/enhancer_map'
 
 #module load python/3.9.2
 python3 --version
-source ./../venv/bin/activate
+source venv/bin/activate
 
 echo "python3 add_annotations.py -f $in_file -p $panther_dir -e $enhancer_dir > $out_file"
 echo start at `date`    
-    python3 add_annotations.py -f $in_file -p  $panther_dir -e $enhancer_dir > $out_file
+python3 -m wgsa_add.add_annotations -f $in_file -p  $panther_dir -e $enhancer_dir > $out_file
 echo end at `date`
